requirements.txt
fastapi==0.114.2
Jinja2==3.1.4
pydantic==2.9.1
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyJWT==2.6.0
uvicorn
passlib[bcrypt]